microservice/oach_cti/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
django-requestlogs==0.2.3
django-ipware==3.0.2
python-dateutil==2.8.1
requests==2.25.1
django-redis==4.12.1
django-eventstream==4.1.0
channels==3.0.3